+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## Change Log/Revision History
 
 
+4.5.1 - Reference Platform Clinical Document Library (xxx 2018 in dev)
+----------------------------------------------------
+- Fix up logo path when specifying different OutputFolder
+- PCML - added identifier for Participant, plus example custom Narrative (not yet supported though)
 
 4.5.0 - Reference Platform Clinical Document Library (Aug 2018)
 ----------------------------------------------------